+++ Changelog.md
@@ -30,3 +30,9 @@ Fix: pip installing python client now also installs `requests` dependencies, whi
 -----
 
 * Python Client: `Client` can now be pickeled again.
+
+Next
+----
+
+* Fairness
+* Large semaphores don't starve